refactor: 당첨번호와의 일치 개수 메소드 - LottoTicket 컬렉션 변경되지 않도록 수정

+++ src/main/java/lotto/domain/LottoRank.java
@@ -20,7 +20,7 @@ public enum LottoRank {
         this.prize = prize;
     }
 
-    public static LottoRank of(int matchCount) {
+    public static LottoRank of(long matchCount) {
         return Stream.of(LottoRank.values())
                 .filter(lottoRank -> lottoRank.matchCount == matchCount)
                 .findFirst()
